fs : use same gid_t, uid_t, ino_t for on disk and generic inodes

+++ kernel/include/stddef.h
@@ -14,8 +14,8 @@ typedef unsigned int time_t;
 typedef int pid_t;
 typedef unsigned short dev_t;
 typedef uint16_t ino_t;
-typedef uint32_t uid_t;
-typedef uint32_t gid_t;
+typedef uint16_t uid_t;
+typedef uint16_t gid_t;
 
 #define NULL                                  ((void *) 0)
 #define UNUSED(x)                             ((void) (x))
