+++ libc/include/sys/types.h
@@ -16,7 +16,7 @@ typedef unsigned int off_t;
 typedef unsigned int time_t;
 typedef int pid_t;
 typedef uint16_t ino_t;
-typedef uint32_t uid_t;
-typedef uint32_t gid_t;
+typedef uint16_t uid_t;
+typedef uint16_t gid_t;
 
 #endif
